mm: Introduce AS_INACCESSIBLE for encrypted/confidential memory
commit c72ceafbd12cf95e088681ae5e535ef1a78bf0ed upstream.

filemap users like guest_memfd may use page cache pages to
allocate/manage memory that is only intended to be accessed by guests
via hardware protections like encryption. Writes to memory of this sort
in common paths like truncation may cause unexpected behavior such as
writing garbage instead of zeros when attempting to zero pages, or
worse, triggering hardware protections that are considered fatal as far
as the kernel is concerned.

Introduce a new address_space flag, AS_INACCESSIBLE, and use this
initially to prevent zero'ing of pages during truncation, with the
understanding that it is up to the owner of the mapping to handle this
specially if needed.

This is admittedly a rather blunt solution, but it seems like
there are no other places that should take into account the
flag to keep its promise.

[Backport Changes]

1. In file include/linux/pagemap.h enum 'mapping_flags' had already been
modified to contain AS_INACCESSIBLE by openeuler commit e6ebbdef6b5db01872e4c5866d333cb2f923be57.
The mapping flags enum had already been updated to a newer kernel
v 6.12-rc1 and above. To maintain conistency existing euler source and
to align source with the future to avoid conflict changing 'mapping_flags'
is skipped.

Link: https://lore.kernel.org/lkml/ZR9LYhpxTaTk6PJX@google.com/
Cc: Matthew Wilcox <willy@infradead.org>
Suggested-by: Sean Christopherson <seanjc@google.com>
Signed-off-by: Michael Roth <michael.roth@amd.com>
Message-ID: <20240329212444.395559-5-michael.roth@amd.com>
Acked-by: Vlastimil Babka <vbabka@suse.cz>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>